+++ superwall/src/main/java/com/superwall/sdk/network/Network.kt
@@ -167,7 +167,7 @@ open class Network(
         }
     }
 
-    suspend fun getAssignments(): List<Assignment> {
+    open suspend fun getAssignments(): List<Assignment> {
         return try {
             val result = urlSession.request(
                 Endpoint.assignments(factory = factory)
@@ -183,6 +183,4 @@ open class Network(
             throw error
         }
     }
-
-
 }