+++ superwall/src/main/java/com/superwall/sdk/network/device/DeviceHelper.kt
@@ -115,6 +115,7 @@ class DeviceHelper(
     val vendorId: String
         get() = Settings.Secure.getString(context.contentResolver, Settings.Secure.ANDROID_ID)
 
+    var platformWrapper: String = ""
 
     private val _locale: Locale = Locale.getDefault()
 
@@ -122,7 +123,7 @@ class DeviceHelper(
         get() {
             return try {
                 Currency.getInstance(_locale)
-            } catch (e: Exception) {
+            } catch (e: Throwable) {
                 null
             }
         }