+++ superwall/src/main/java/com/superwall/sdk/paywall/presentation/PublicPresentation.kt
@@ -137,7 +137,7 @@ private suspend fun Superwall.trackAndPresentPaywall(
 ) {
     try {
         TrackingLogic.checkNotSuperwallEvent(event)
-    } catch (e: Exception) {
+    } catch (e: Throwable) {
         return
     }